# Application/src/main/resources/swagger.yaml
swagger: "2.0"
info:
  title: Discount Hunter API
  description: "This is a specification for Discount Hunter API which allows to control whichweb pages
  are monitored for changing prices."
  version: 1.0.0
host: "localhost:8080"
basePath: /
tags:
  - name: "Web Pages"
    description: "Everything about Web Pages in the system"
schemes:
  - http
paths:
  /webPages/{id}:
    get:
      summary: "Retrieve existing web page information along with its price details"
      operationId: "getWebPage"
      produces:
        - application/json
      parameters:
        - in: "path"
          name: "id"
          description: "Id of webpage to be retrieved"
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: "OK"
          schema:
            $ref: '#/definitions/WebPageDto'
  /webPages:
    post:
      summary: "Add a new web page to web page collection"
      operationId: "addWebPage"
      produces:
        - application/json
      parameters:
        - in: "body"
          name: "WebPageDto"
          description: "Web Page Object that should be added"
          required: true
          schema:
            $ref: '#/definitions/WebPageDto'
      responses:
        201:
          description: "Web Page Created"
          schema:
            $ref: '#/definitions/WebPageDto'
  /webPages/priceDetails/{webPageId}:
    get:
      summary: "Get all price details for specific web page"
      operationId: "getPriceDetails"
      produces:
        - application/json
      parameters:
        - in: "path"
          name: "webPageId"
          required: true
          type: integer
          format: int64
      responses:
        200:
          description: "Price details retrieved successfully"
          schema:
            type: "array"
            items:
              $ref: '#/definitions/WebPagePriceDetailsDto'
definitions:
  WebPageDto:
    type: object
    required:
      - url
    properties:
      id:
        type: integer
        format: int64
      url:
        type: string
    xml:
      name: DoctorDto
  WebPagePriceDetailsDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      pageId:
        type: integer
        format: int64
      price:
        type: string